FIxed #978

+++ Ombi.Api.Models/Radarr/RadarrMovieResponse.cs
@@ -47,7 +47,7 @@ namespace Ombi.Api.Models.Radarr
     {
         public string title { get; set; }
         public string sortTitle { get; set; }
-        public int sizeOnDisk { get; set; }
+        public double sizeOnDisk { get; set; }
         public string status { get; set; }
         public string overview { get; set; }
         public string inCinemas { get; set; }
